Revert "Don't use uint8_t type."
This reverts commit 8eb7e7c7ba8c5d3d9c5ea74189c6213143d10b95.

+++ libhotp/hotp.c
@@ -232,7 +232,7 @@ hotp_generate_otp (const char *secret,
   }
 
   {
-    size_t offset = hs[sizeof (hs) - 1] & 0x0f;
+    uint8_t offset = hs[sizeof (hs) - 1] & 0x0f;
 
     S = (((hs[offset] & 0x7f) << 24)
 	 | ((hs[offset + 1] & 0xff) << 16)
